+++ internal/resources/cloudintegrations/models.go
@@ -1,4 +1,4 @@
-package integrations
+package cloudintegrations
 
 import "time"
