+++ internal/resources/cloudintegrations/resource_integration.go
@@ -1,4 +1,4 @@
-package integrations
+package cloudintegrations
 
 import (
 	"context"
