+++ pkg/provider/resources.go
@@ -7,12 +7,12 @@ import (
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/appplatform"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/asserts"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/cloud"
+	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/cloudintegrations"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/cloudprovider"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/connections"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/fleetmanagement"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/frontendo11y"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/grafana"
-	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/integrations"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/k6"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/machinelearning"
 	"github.com/grafana/terraform-provider-grafana/v4/internal/resources/oncall"
@@ -68,7 +68,7 @@ func Resources() []*common.Resource {
 	var resources []*common.Resource
 	resources = append(resources, cloud.Resources...)
 	resources = append(resources, grafana.Resources...)
-	resources = append(resources, integrations.Resources...)
+	resources = append(resources, cloudintegrations.Resources...)
 	resources = append(resources, oncall.Resources...)
 	resources = append(resources, machinelearning.Resources...)
 	resources = append(resources, slo.Resources...)
